Promote MockURLSession to ApolloTestSupport instead of a new mock
MockApolloURLSession duplicated machinery the dev repo already had, and
did it with lower fidelity: it required callers to pre-split multi-part
chunks, bypassing AsyncHTTPResponseChunkSequence's real boundary
parsing, and its ChunkSequence duplicated MockAsyncChunkSequence.

Ship the existing, proven helpers instead. MockResponseProvider and
MockURLProtocol move into ApolloTestSupport verbatim, joined by an
HTTP-only MockURLSession. Because responses are served through a
URLProtocol backed by a real URLSession, multi-part bodies are chunked
by the same code path a real response uses.

The WebSocket half stays in the dev repo as MockWebSocketURLSession so
ApolloTestSupport does not need to depend on ApolloWebSocket. Only
WebSocketTests used it, and WebSocketTransport requires just
WebSocketURLSession, so the split needed no other changes.

Test files that used these helpers now import ApolloTestSupport
directly.
