+++ Tests/ApolloTests/Cache/LoadQueryFromStoreTests.swift
@@ -1,5 +1,6 @@
 @_spi(Execution) @_spi(Unsafe) import ApolloAPI
 @_spi(Execution) import ApolloInternalTestHelpers
+import ApolloTestSupport
 import Nimble
 import XCTest
