+++ Tests/ApolloTests/Interceptors/MaxRetryInterceptorTests.swift
@@ -1,5 +1,6 @@
 import ApolloAPI
 import ApolloInternalTestHelpers
+import ApolloTestSupport
 import Nimble
 import XCTest
