+++ Tests/ApolloTests/Interceptors/ResponseCodeInterceptorTests.swift
@@ -1,6 +1,7 @@
 @_spi(Internal) import Apollo
 import ApolloAPI
 import ApolloInternalTestHelpers
+import ApolloTestSupport
 import Nimble
 import XCTest
