+++ Tests/ApolloTests/Network/ApolloURLSessionTests.swift
@@ -1,4 +1,5 @@
 import ApolloInternalTestHelpers
+import ApolloTestSupport
 import Nimble
 import XCTest
